Split main thread code in separate files
