+++ src/mainThread/sessionOptions.ts
@@ -1,0 +1,4 @@
+export class SessionOptions
+{
+    public showClearDataDialog: boolean = true;
+}
